refactor(sql): split gnrsqltable.py into class-based sub-package
Replace the monolithic gnrsqltable.py (~3200 lines) with a focused
sub-package using the mixin pattern (same approach as gnrsql/ and
gnrsqlmodel/).

New package structure:
- __init__.py: facade with exceptions, EXCEPTIONS dict, re-exports
- helpers.py: add_sql_comment, orm_audit_log decorators, RecordUpdater
- columns.py: ColumnsMixin — column access, properties, variant columns
- query.py: QueryMixin — core query(), WHERE translation, column utils
- record.py: RecordMixin — record building, caching, captions
- crud.py: CrudMixin — insert/update/delete, batch, writeRecordCluster
- triggers.py: TriggersMixin — hooks, protection, validation, events
- serialization.py: SerializationMixin — JSON/XML import/export
- copy.py: CopyMixin — copy/paste, duplication, archival
- utils.py: UtilsMixin — pkey generation, data export, totalizers
- table.py: SqlTable assembles all mixins + GnrObject

All existing imports continue to work unchanged via __init__.py.
REVIEW markers added for 4 confirmed bugs and 11 other issues.

Ref #498
